Refactoring configuration
