+++ samples/apps/copilot-chat-app/webapi/Config/AzureSpeechOptions.cs
@@ -2,11 +2,20 @@
 
 namespace SemanticKernel.Service.Config;
 
+/// <summary>
+/// Configuration options for Azure speech recognition.
+/// </summary>
 public sealed class AzureSpeechOptions
 {
     public const string PropertyName = "AzureSpeech";
 
+    /// <summary>
+    /// Region of the Azure speech service.
+    /// </summary>
     public string Region { get; set; } = string.Empty;
 
+    /// <summary>
+    /// Key to access the Azure speech service.
+    /// </summary>
     public string Key { get; set; } = string.Empty;
 }
