+++ samples/apps/copilot-chat-app/webapi/ConfigurationExtensions.cs
@@ -1,0 +1,47 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System.Reflection;
+using Azure.Identity;
+
+namespace SemanticKernel.Service;
+
+internal static class ConfigExtensions
+{
+    public static IHostBuilder AddConfiguration(this IHostBuilder host)
+    {
+        string? environment = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
+
+        host.ConfigureAppConfiguration((builderContext, configBuilder) =>
+        {
+            configBuilder.AddJsonFile(
+                path: "appsettings.json",
+                optional: false,
+                reloadOnChange: true);
+
+            configBuilder.AddJsonFile(
+                path: $"appsettings.{environment}.json",
+                optional: true,
+                reloadOnChange: true);
+
+            configBuilder.AddEnvironmentVariables();
+
+            configBuilder.AddUserSecrets(
+                assembly: Assembly.GetExecutingAssembly(),
+                optional: true,
+                reloadOnChange: true);
+
+            // For settings from Key Vault, see https://learn.microsoft.com/en-us/aspnet/core/security/key-vault-configuration?view=aspnetcore-8.0
+            string? keyVaultUri = builderContext.Configuration["KeyVaultUri"];
+            if (!string.IsNullOrWhiteSpace(keyVaultUri))
+            {
+                configBuilder.AddAzureKeyVault(
+                    new Uri(keyVaultUri),
+                    new DefaultAzureCredential());
+
+                // for more information on how to use DefaultAzureCredential, see https://learn.microsoft.com/en-us/dotnet/api/azure.identity.defaultazurecredential?view=azure-dotnet
+            }
+        });
+
+        return host;
+    }
+}
